Support responses with custom MIME type and headers
Support responses with a custom MIME type and headers by using the
decorator @use_custom_headers on a view function and by having the view
function return a dict of "content", "mimetype" and "headers",
specifying the actual content, MIME type and possible additional
headers, respectively. The decorator sets the attribute
use_custom_headers on the view function. When main_handler notices the
attribute, it calls the nested function make_custom_response instead of
using incremental_json or full_json.

This feature makes it possible to add endpoints returning non-JSON
responses, for example, for downloading data in a specific format.